chore(test): refactor tests
